# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PubKeyMail API
  version: 1.0.0
  description: >-
    Blockchain-based email service with wallet authentication.


    ## Authentication

    PubKeyMail uses wallet-based authentication. Users sign a challenge message 

    with their blockchain wallet and receive a JWT token for subsequent
    requests.


    ## Case Sensitivity

    **CRITICAL**: All blockchain addresses are case-sensitive. Different casing 

    represents different addresses.
  contact:
    name: PubKeyMail Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.pubkeymail.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/verify endpoint
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request parameters
            details:
              type: array
              items: {}
          required:
            - code
            - message
      required:
        - error
...
    RegisterRequest:
      type: object
      properties:
        address:
          type: string
          minLength: 1
          example: GNa9E2dWP8e2fg23pRJoiCvJm5cbzomm94YzotpcRh7A
        blockchain:
          type: string
          enum:
            - solana
            - ethereum
            - polygon
          default: solana
      required:
        - address
    LinkedAddress:
      type: object
      properties:
        id:
          type: number
        address:
          type: string
        blockchain:
          type: string
          enum:
            - solana
            - ethereum
            - polygon
        verifiedAt:
          type: string
          format: date-time
      required:
        - address
        - blockchain
    User:
      type: object
      properties:
        id:
          type: number
          example: 1
        primaryAddress:
          type: string
          example: GNa9E2dWP8e2fg23pRJoiCvJm5cbzomm94YzotpcRh7A
        blockchain:
          type: string
          enum:
            - solana
            - ethereum
            - polygon
        subscriptionStatus:
          type: string
          enum:
            - inactive
            - active
            - cancelled
            - expired
        subscriptionTier:
          type: string
          enum:
            - free
            - paid
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - primaryAddress
        - blockchain
        - subscriptionStatus
        - subscriptionTier
        - createdAt
    UserWithAddresses:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            linkedAddresses:
              type: array
              items:
                $ref: "#/components/schemas/LinkedAddress"
          required:
            - linkedAddresses
    LinkAddressRequest:
      type: object
      properties:
        address:
          type: string
          minLength: 1
        blockchain:
          type: string
          enum:
            - solana
            - ethereum
            - polygon
          default: solana
      required:
        - address
...
paths:
...
  /api/v1/users/register:
    post:
      tags:
        - Users
      summary: Register new user
      description: Register a new user with a wallet address
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                required:
                  - user
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/users/check/{address}:
    get:
      tags:
        - Users
      summary: Check if address is registered
      parameters:
        - schema:
            type: string
            description: Blockchain wallet address
          required: true
          description: Blockchain wallet address
          name: address
          in: path
      responses:
        "200":
          description: Registration status
          content:
            application/json:
              schema:
                type: object
                properties:
                  registered:
                    type: boolean
                required:
                  - registered
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/users/profile/{address}:
    get:
      tags:
        - Users
      summary: Get user profile by address
      parameters:
        - schema:
            type: string
            description: Blockchain wallet address
          required: true
          description: Blockchain wallet address
          name: address
          in: path
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/UserWithAddresses"
                required:
                  - user
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/users/{userId}/link-address:
    post:
      tags:
        - Users
      summary: Link additional address to user
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type:
              - number
              - "null"
            description: User ID
          required: false
          description: User ID
          name: userId
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkAddressRequest"
      responses:
        "201":
          description: Address linked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  linkedAddress:
                    $ref: "#/components/schemas/LinkedAddress"
                required:
                  - linkedAddress
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Address already linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/users/{userId}/addresses/{addressId}:
    delete:
      tags:
        - Users
      summary: Unlink address from user
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type:
              - number
              - "null"
            description: User ID
          required: false
          description: User ID
          name: userId
          in: path
        - schema:
            type:
              - number
              - "null"
            description: Address ID
          required: false
          description: Address ID
          name: addressId
          in: path
      responses:
        "200":
          description: Address unlinked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Address unlinked successfully
                required:
                  - message
        "400":
          description: Cannot unlink primary address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Address not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...